+++ cmd/crank/xpkg/logout.go
@@ -26,7 +26,7 @@ import (
 
 	"github.com/crossplane/crossplane-runtime/pkg/errors"
 
-	"github.com/crossplane/crossplane/cmd/crank/internal/upbound"
+	"github.com/crossplane/crossplane/internal/xpkg/upbound"
 )
 
 const (
