+++ internal/xpkg/upbound/context.go
@@ -31,7 +31,7 @@ import (
 
 	"github.com/crossplane/crossplane-runtime/pkg/errors"
 
-	"github.com/crossplane/crossplane/cmd/crank/internal/upbound/config"
+	"github.com/crossplane/crossplane/internal/xpkg/upbound/config"
 )
 
 const (
