+++ internal/xpkg/upbound/context_test.go
@@ -30,7 +30,7 @@ import (
 	"github.com/crossplane/crossplane-runtime/pkg/errors"
 	"github.com/crossplane/crossplane-runtime/pkg/test"
 
-	"github.com/crossplane/crossplane/cmd/crank/internal/upbound/config"
+	"github.com/crossplane/crossplane/internal/xpkg/upbound/config"
 )
 
 var (
