+++ internal/xpkg/upbound/credhelper/credhelper.go
@@ -25,7 +25,7 @@ import (
 	"github.com/crossplane/crossplane-runtime/pkg/errors"
 	"github.com/crossplane/crossplane-runtime/pkg/logging"
 
-	"github.com/crossplane/crossplane/cmd/crank/internal/upbound/config"
+	"github.com/crossplane/crossplane/internal/xpkg/upbound/config"
 )
 
 const (
