+++ internal/xpkg/upbound/credhelper/credhelper_test.go
@@ -25,7 +25,7 @@ import (
 	"github.com/crossplane/crossplane-runtime/pkg/errors"
 	"github.com/crossplane/crossplane-runtime/pkg/test"
 
-	"github.com/crossplane/crossplane/cmd/crank/internal/upbound/config"
+	"github.com/crossplane/crossplane/internal/xpkg/upbound/config"
 )
 
 // TODO(hasheddan): these tests are testing through to the underlying config
